backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
prophet==1.1.5
